Fix some python FOCS errors
