+++ default/scripting/focs/_industry.pyi
@@ -1,0 +1,6 @@
+from focs._types import _EffectGroup
+
+def _industry(tag, default_multiplier) -> list[_EffectGroup]:  # type: ignore[reportGeneralTypeIssues]
+    ...
+
+_BASIC_INDUSTRY: list[_EffectGroup]
