+++ test-scripting/focs/_buildings.py
@@ -1,0 +1,1 @@
+# Placeholder for module.
