+++ test-scripting/focs/_tech.py
@@ -1,0 +1,1 @@
+# Placeholder for module.
